backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic[email]
pydantic-settings==2.0.3
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
asyncpg==0.28.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.0.0
pytest==7.4.2
httpx==0.25.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
aiosqlite==0.19.0
black==23.7.0
ruff==0.0.291
